+++ smarty-woocommerce-checkout-progress-bar.php
@@ -828,3 +828,10 @@ if (!function_exists('smarty_cpb_load_changelog')) {
     }
     add_action('wp_ajax_smarty_cpb_load_changelog', 'smarty_cpb_load_changelog');
 }
+
+// Add a links on the Plugins page
+add_filter('plugin_action_links_' . plugin_basename(__FILE__), function($links) {
+    $links[] = '<a href="' . admin_url('admin.php?page=smarty-cpb-settings') . '">' . __('Settings', 'smarty-woocommerce-checkout-progress-bar') . '</a>';
+    $links[] = '<a href="https://github.com/mnestorov/smarty-woocommerce-checkout-progress-bar" target="_blank">' . __('GitHub', 'smarty-woocommerce-checkout-progress-bar') . '</a>';
+    return $links;
+});