+++ CHANGELOG.md
@@ -14,4 +14,5 @@
 ### 1.0.2 (2025.06.04)
 - Threshold labels are now editable from the settings page
 - Move the admin css from the main file to the separate css file: `smarty-cpb-admin.css`
+- Add a links on the Plugins page
 - Bug fixes